Renombrado el fichero que hace la conexion a la base de datos
